Modified 5/6/2016
*modified Display subprogram, now displays correctly, has not checked for
bugs and have not checked to see which color the piece changes to when on
water. Still need to modify the flag display and the trap display

*added small subprogram printPlayer in order to avoid repetition, should
not be a problem

*still a few bugs that have not been modified, still need to debug and
fix.

*hardest part which is moving the pieces still has not been implemented,
will have to do it after exams.

*My other teammates still have not done anything, still waiting for them
to do something, if they have not done anything, i will notify the teacher
and make sure my hard work will be noticed.

END OF COMMENTS 5/6/2016